requirements.txt
langchain==0.0.341
python-dotenv==1.0.0
Requests==2.31.0
streamlit==1.28.2
